Combine control bar & time progress bar into one thing
